requirements.txt
pandas
streamlit
streamlit-extras
snowflake-sqlalchemy
python-dotenv
sqlalchemy<2.0
snowflake-connector-python


# sqlalchemy
# snowflake-connector-python==3.0.2
# pyarrow==10.0.1
# python-dotenv==1.0.0
# snowflake-connector-python[pandas]==1.16.0
# snowflake-sqlalchemy
